+++ CHANGELOG.md
@@ -41,6 +41,8 @@ Other guiding principles:
 ### Added
 
 - **amaru-consensus**: track peer-sharing reputation (handshake advertisability, connection failures, sticky adversarial flag, and whether a successful connection was ever established) so peer selection can filter share candidates. ([#1167](https://github.com/pragma-org/amaru/issues/1167))
+- **amaru-protocols**: add peer-sharing mini-protocol and initiator ([#1168](https://github.com/pragma-org/amaru/issues/1168))
+- **amaru-consensus**: used shared peers to populate peer candidate pool ([#1168](https://github.com/pragma-org/amaru/issues/1168))
 
 ### Fixed